aloy ascension refactor
